Add pay wall (#3)
* Added Laravel Cashier and started implementing a paywall.

* Ran pint and fixed some issues that were found.

* Added timezone to the created_at attribute.

* Added the ability to checkout via stripe.

* Added some methods to the user model to help with billing.
Added some pages to display to the user when they need to upgrade.

* Made some small improvements.

* Opened up registration.

+++ app/Http/Controllers/StudentController.php
@@ -30,8 +30,11 @@ class StudentController extends Controller
     {
         return Inertia::render('Teachers/Students/Index', [
             'students' => $request->user()->students()->withCount(['promptQuestions' => function (Builder $query) {
-                $query->whereHas('promptAnswer')->filterByDate(today()->toDateString());
-            }])->paginate(10),
+                $query->whereHas('promptAnswer');
+            }])
+                ->paginate(10),
+            'showInitialPaymentPage' => $request->user()->showInitialPaymentPage(),
+            'showExceededQuantityPage' => $request->user()->showExceededQuantityPage(),
         ]);
     }
 
@@ -69,6 +72,8 @@ class StudentController extends Controller
 
     public function show(Request $request, Student $student): Response
     {
+        $this->authorize('view', $student);
+
         return Inertia::render('Teachers/Students/Show', [
             'student' => (new StudentResource($student->load('promptQuestions')))->resolve(),
             'totalQuestions' => $this->studentService->student($student)->totalQuestionsAsked(),
@@ -80,6 +85,8 @@ class StudentController extends Controller
 
     public function update(StudentStoreRequest $request, Student $student): RedirectResponse
     {
+        $this->authorize('update', $student);
+
         $student->update($request->validated());
 
         return to_route('students.index');
@@ -87,6 +94,8 @@ class StudentController extends Controller
 
     public function destroy(Student $student): RedirectResponse
     {
+        $this->authorize('update', $student);
+
         $student->delete();
 
         return to_route('students.index');
